70% Assignment Completion
Made enough progress to secure a grade of 70% if it was submitted now.

Added the functionality of only showing book's that the current signed in user has read.
Added bootstrap and started to add design to the site using the added framework.
Removed all old CSS to accommodate bootstrap.
Changed the forms and data structure from generic people information to the books per user account.
Added the mongoDB connection to pull from the external database.
Added a nav bar with a logo (WPI).