Added RandomNumber Generator
Need to fix the error of populating on the frontend only the recipes for a particular user rather than showing all recipes. I am unable to do for a particular user because I do not have the user email. How to get user email from userModel? How to use cookies to store user credentials temporary?